requirement.txt
asgiref==3.8.1
certifi==2025.1.31
Django==5.1.6
django-filter==24.3
influxdb-client==1.48.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
reactivex==4.0.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
